+++ apps/rig_api/lib/rig_api/endpoint.ex
@@ -29,6 +29,11 @@ defmodule RigApi.Endpoint do
     signing_salt: "7t9/VVWp"
   )
 
+  # Prometheus Integration - START
+  # makes the /metrics URL happen
+  plug(RigMetrics.MetricsPlugExporter)
+  # Prometheus Integration - END
+
   plug(RigApi.Router)
 
   def init(_key, config) do